Refactor Timetable, TimetableSnapshot, TripPattern is moved from the edge to the model package.
