+++ src/main/java/org/opentripplanner/model/TimetableSnapshotProvider.java
@@ -1,7 +1,5 @@
 package org.opentripplanner.model;
 
-import org.opentripplanner.routing.edgetype.TimetableSnapshot;
-
 /**
  * This interface is used to retrieve the current instance of the TimetableSnapshot. Any provider
  * implementing this interface is responsible for thread-safe access to the latest valid
