+++ src/main/java/org/opentripplanner/updater/stoptime/TripPatternCache.java
@@ -6,7 +6,7 @@ import java.util.Map;
 import org.opentripplanner.model.FeedScopedId;
 import org.opentripplanner.model.Route;
 import org.opentripplanner.model.StopPattern;
-import org.opentripplanner.routing.edgetype.TripPattern;
+import org.opentripplanner.model.TripPattern;
 import org.opentripplanner.routing.graph.Graph;
 
 /**
