+++ src/test/java/org/opentripplanner/netex/loader/mapping/TripPatternMapperTest.java
@@ -4,7 +4,7 @@ import org.junit.Test;
 import org.opentripplanner.model.Stop;
 import org.opentripplanner.model.Trip;
 import org.opentripplanner.netex.loader.util.HierarchicalMapById;
-import org.opentripplanner.routing.edgetype.TripPattern;
+import org.opentripplanner.model.TripPattern;
 import org.opentripplanner.routing.trippattern.Deduplicator;
 import org.opentripplanner.routing.trippattern.TripTimes;
